+++ handlers/issue/sum-total-scores-per-contributor.ts
